+++ pkg/quarterdeck/orgs.go
@@ -6,15 +6,11 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/golang-jwt/jwt/v4"
 	"github.com/oklog/ulid/v2"
 	"github.com/rotationalio/ensign/pkg/quarterdeck/api/v1"
 	"github.com/rotationalio/ensign/pkg/quarterdeck/db/models"
 	"github.com/rotationalio/ensign/pkg/quarterdeck/middleware"
-	"github.com/rotationalio/ensign/pkg/quarterdeck/permissions"
-	"github.com/rotationalio/ensign/pkg/quarterdeck/responses"
 	"github.com/rotationalio/ensign/pkg/quarterdeck/tokens"
-	"github.com/rotationalio/ensign/pkg/utils/metrics"
 	"github.com/rotationalio/ensign/pkg/utils/pagination"
 	"github.com/rotationalio/ensign/pkg/utils/sentry"
 	"github.com/rotationalio/ensign/pkg/utils/ulids"
@@ -148,139 +144,4 @@ func (s *Server) OrganizationDetail(c *gin.Context) {
 
 	// Build the response from the model
 	c.JSON(http.StatusOK, org.ToAPI())
-}
-
-func (s *Server) ProjectCreate(c *gin.Context) {
-	var (
-		err     error
-		project *api.Project
-		claims  *tokens.Claims
-	)
-
-	// Bind the Project request to the project data structure
-	if err = c.BindJSON(&project); err != nil {
-		sentry.Warn(c).Err(err).Msg("could not parse project create request")
-		c.JSON(http.StatusBadRequest, api.ErrorResponse(responses.ErrTryProjectAgain))
-		return
-	}
-
-	// Validate the request from the API side.
-	if err = project.Validate(); err != nil {
-		c.Error(err)
-		c.JSON(http.StatusBadRequest, api.ErrorResponse(responses.ErrFixProjectDetails))
-		return
-	}
-
-	// Fetch the user claims from the request
-	if claims, err = middleware.GetClaims(c); err != nil {
-		sentry.Error(c).Err(err).Msg("could not get user claims from authenticated request")
-		c.JSON(http.StatusInternalServerError, api.ErrorResponse(responses.ErrSomethingWentWrong))
-		return
-	}
-
-	// Create the OrganizationProjects mapping database model
-	model := &models.OrganizationProject{
-		OrgID:     claims.ParseOrgID(),
-		ProjectID: project.ProjectID,
-	}
-
-	// Save the model to the database
-	if err = model.Save(c.Request.Context()); err != nil {
-		switch err.(type) {
-		case *models.ValidationError:
-			c.Error(err)
-			c.JSON(http.StatusBadRequest, api.ErrorResponse(responses.ErrTryProjectAgain))
-		default:
-			sentry.Error(c).Err(err).Msg("could not create project")
-			c.JSON(http.StatusInternalServerError, api.ErrorResponse(responses.ErrSomethingWentWrong))
-		}
-		return
-	}
-
-	// Update the response to send to the user
-	project.OrgID = model.OrgID
-	project.ProjectID = model.ProjectID
-	project.Created, _ = model.GetCreated()
-	project.Modified, _ = model.GetModified()
-
-	// Increment total number of projects in prometheus
-	metrics.Projects.WithLabelValues(ServiceName).Inc()
-	c.JSON(http.StatusOK, project)
-}
-
-func (s *Server) ProjectAccess(c *gin.Context) {
-	var (
-		err     error
-		exists  bool
-		project *api.Project
-		claims  *tokens.Claims
-		creds   *api.LoginReply
-	)
-
-	// Bind the Project request to the project data structure
-	if err = c.BindJSON(&project); err != nil {
-		sentry.Warn(c).Err(err).Msg("could not parse project access request")
-		c.JSON(http.StatusBadRequest, api.ErrorResponse(responses.ErrTryProjectAgain))
-		return
-	}
-
-	// Validate the request from the API side.
-	if err = project.Validate(); err != nil {
-		c.Error(err)
-		c.JSON(http.StatusBadRequest, api.ErrorResponse(err))
-		return
-	}
-
-	// Fetch the user claims from the request
-	if claims, err = middleware.GetClaims(c); err != nil {
-		sentry.Error(c).Err(err).Msg("could not get user claims from authenticated request")
-		c.JSON(http.StatusInternalServerError, api.ErrorResponse(responses.ErrSomethingWentWrong))
-		return
-	}
-
-	// Check that the orgID claim is allowed access to the project
-	model := &models.OrganizationProject{OrgID: claims.ParseOrgID(), ProjectID: project.ProjectID}
-	if exists, err = model.Exists(c.Request.Context()); !exists || err != nil {
-		if !exists {
-			c.JSON(http.StatusBadRequest, api.ErrorResponse(responses.ErrTryProjectAgain))
-			return
-		}
-
-		sentry.Error(c).Err(err).Msg("could not retrieve organization project mapping from database")
-		c.JSON(http.StatusInternalServerError, api.ErrorResponse(responses.ErrSomethingWentWrong))
-		return
-	}
-
-	// Prepare the credentials for the access token to make the one time access.
-	now := time.Now()
-	ota := &tokens.Claims{
-		RegisteredClaims: jwt.RegisteredClaims{
-			ID:        ulids.New().String(),
-			Subject:   claims.Subject,
-			IssuedAt:  jwt.NewNumericDate(now),
-			NotBefore: jwt.NewNumericDate(now),
-			ExpiresAt: jwt.NewNumericDate(now.Add(OneTimeAccessDuration)),
-		},
-		OrgID:       claims.OrgID,
-		ProjectID:   project.ProjectID.String(),
-		Permissions: make([]string, 0, 4),
-	}
-
-	// Add only the user permissions related to topics to these claims -- whatever
-	// access to topics the user has, so to will the one time access claims.
-	for _, permission := range claims.Permissions {
-		if permissions.InGroup(permission, permissions.PrefixTopics) {
-			ota.Permissions = append(ota.Permissions, permission)
-		}
-	}
-
-	creds = &api.LoginReply{}
-	token := s.tokens.CreateToken(ota)
-	if creds.AccessToken, err = s.tokens.Sign(token); err != nil {
-		sentry.Error(c).Err(err).Msg("could not sign token")
-		c.JSON(http.StatusInternalServerError, api.ErrorResponse(responses.ErrSomethingWentWrong))
-		return
-	}
-
-	c.JSON(http.StatusOK, creds)
 }
