+++ pkg/quarterdeck/responses/errors.go
@@ -7,6 +7,7 @@ var (
 	ErrTryProjectAgain       = "Unable to create or access that project - please try again!"
 	ErrFixProjectDetails     = "Unable to create a project with those details - please correct them and try again!"
 	ErrNeedProjectPermission = "Unable to access project - please request permission from your team owner."
+	ErrProjectNotFound       = "Project with the specified ID was not found."
 	ErrLogBackIn             = "Logged out of your account - please log back in!"
 	ErrVerifyEmail           = "Please verify your email address and try again!"
 	ErrVerificationFailed    = "Email verification failed. Please contact support@rotational.io for assistance."
